Moved Camera Switch & made an AutoFocus
Updated:
- Camera Switch to take optional target (Defaults to self).
  - Added changeable zoom level
- Added Cinemachine to asmdef
- made focal point smaller / recolored to be different from spawn point
