+++ Assets/Prefabs/Camera/FocalPoint.prefab
@@ -44,5 +44,5 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e113d5f475f0548449d2115886cd6286, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _color: {r: 0, g: 0, b: 1, a: 1}
-  _dim: 1
+  _color: {r: 0, g: 1, b: 0.75454235, a: 1}
+  _dim: 0.5
