+++ Assets/Scripts/Camera/AutoCameraFocus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: deb70e044b2616a488c0362dab046db1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
