+++ Assets/Scripts/Camera/CameraSwitch.cs
@@ -1,0 +1,36 @@
+using Cinemachine;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CameraSwitch : MonoBehaviour
+{
+    [SerializeField, Tooltip("Default self")] 
+    private Transform _target;
+
+    [SerializeField, Tooltip("Default 0.5f")]
+    private float _zoomLevel = 0.5f;
+
+    private CinemachineFreeLook _camera;
+
+    private void Start()
+    {
+        if (!_camera)
+        {
+            _camera = FindObjectOfType<CinemachineFreeLook>();
+        }
+
+        if(!_target)
+        {
+            _target = this.transform;
+        }
+    }
+
+    private void OnMouseDown()
+    {
+        _camera.m_YAxis.Value = _zoomLevel;
+        _camera.Follow = _target;
+        _camera.LookAt = _target;
+        
+    }
+}
