+++ Assets/Scripts/Scripts.asmdef
@@ -11,7 +11,8 @@
         "GUID:52087dca63f4bcd4eac3a0df270d946d",
         "GUID:bc7babff56e1452478aeab958c9efba5",
         "GUID:75469ad4d38634e559750d17036d5f7c",
-        "GUID:6055be8ebefd69e48b49212b09b47b2f"
+        "GUID:6055be8ebefd69e48b49212b09b47b2f",
+        "GUID:4307f53044263cf4b835bd812fc161a4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
